public-api/requirements.txt
fastapi[standard]==0.112.0
uvicorn==0.34.2
python-dotenv==1.0.0